move physical expression planning
