+++ datafusion/physical-expr/src/udf.rs
@@ -18,10 +18,10 @@
 //! UDF support
 
 use super::type_coercion::coerce;
-use crate::error::Result;
+use crate::{PhysicalExpr, ScalarFunctionExpr};
 use arrow::datatypes::Schema;
+use datafusion_common::Result;
 pub use datafusion_expr::ScalarUDF;
-use datafusion_physical_expr::{PhysicalExpr, ScalarFunctionExpr};
 use std::sync::Arc;
 
 /// Create a physical expression of the UDF.
